requirements.txt
langchain
huggingface
python-dotenv
langchain_huggingface
langchain_google_genai
pydantic
streamlit
PyPDF2

